rerender optimization and error snackbar added

+++ client/src/components/post/deletePost.component.tsx
@@ -1,38 +1,45 @@
-import React, { useEffect } from "react";
+import React, { useContext } from "react";
 import { useMutation } from "@apollo/react-hooks";
 
+import { AuthContext } from "../../context/authContext";
 import { DeletePostProps } from "../../interface/post.interface";
 import { GQL_DELETE_POST } from "../../constants/post";
 import { postsValue } from "../../context/postContext";
+import { errorValue } from "../../context/errorContext";
 
 import IconButton from "@mui/material/IconButton";
 import DeleteIcon from "@mui/icons-material/DeleteForever";
+import CircularProgress from "@mui/material/CircularProgress/CircularProgress";
 
-const DeletePost = (post: DeletePostProps) => {
-  const [deletePost, { data }] = useMutation(GQL_DELETE_POST, {
-    variables: { postId: post.postId },
-  });
-
-  const handleSubmission = () => {
-    if (window.confirm("Are you sure to delete this post?") === true)
-      deletePost();
-  };
-
-  useEffect(() => {
-    if (data?.deletePost === "Post deleted successfully") {
+const DeletePost = ({ postId, userName }: DeletePostProps) => {
+  const { user } = useContext(AuthContext);
+  const [deletePost, { loading }] = useMutation(GQL_DELETE_POST, {
+    variables: { postId: postId },
+    onCompleted: () => {
       const index = postsValue()
-        .map(post => post?.id)
-        .indexOf(post?.postId);
+        .map(post => post.id)
+        .indexOf(postId);
 
       // deep copy the posts value
       const posts = [...postsValue()];
       posts?.splice(index, 1);
 
       postsValue(posts);
-    }
-  }, [data, post]);
+    },
+    onError: error => {
+      errorValue(error);
+    },
+  });
 
-  if (post.currentUser === post.userName) {
+  const handleSubmission = () => {
+    if (window.confirm("Are you sure to delete this ") === true) deletePost();
+  };
+
+  if (loading) {
+    return <CircularProgress style={{ marginLeft: "50%" }} />;
+  }
+
+  if (user.userName === userName) {
     return (
       <IconButton onClick={handleSubmission}>
         <DeleteIcon />
